+++ services/kusto/mgmt/2019-01-21/kusto/models.go
@@ -702,7 +702,7 @@ type DatabaseProperties struct {
 	SoftDeletePeriod *string `json:"softDeletePeriod,omitempty"`
 	// HotCachePeriod - The time the data that should be kept in cache for fast queries in TimeSpan.
 	HotCachePeriod *string `json:"hotCachePeriod,omitempty"`
-	// Statistics - The statistics of the database.
+	// Statistics - READ-ONLY; The statistics of the database.
 	Statistics *DatabaseStatistics `json:"statistics,omitempty"`
 }
 
@@ -714,9 +714,6 @@ func (dp DatabaseProperties) MarshalJSON() ([]byte, error) {
 	}
 	if dp.HotCachePeriod != nil {
 		objectMap["hotCachePeriod"] = dp.HotCachePeriod
-	}
-	if dp.Statistics != nil {
-		objectMap["statistics"] = dp.Statistics
 	}
 	return json.Marshal(objectMap)
 }