+++ services/kusto/mgmt/2020-09-18/kusto/CHANGELOG.md
@@ -1,2 +1,9 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. EventGridConnectionProperties.MarshalJSON() ([]byte, error)
+1. EventHubConnectionProperties.MarshalJSON() ([]byte, error)
+1. IotHubConnectionProperties.MarshalJSON() ([]byte, error)