+++ services/kusto/mgmt/2021-01-01/kusto/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. EventGridConnectionProperties.MarshalJSON() ([]byte, error)
+1. EventHubConnectionProperties.MarshalJSON() ([]byte, error)
+1. IotHubConnectionProperties.MarshalJSON() ([]byte, error)
+1. ScriptProperties.MarshalJSON() ([]byte, error)